Limit pagination results to posts
The list of posts from the homepage included posts and pages, rather
than just all posts as I would have prefered (choosing to link pages
manually).

This change means that non-post pages will not be displayed in the list.
Note however that this is just for the display; the paginator will still
include the pages when calculating how many pages the paginator should
have for example.

Rather than hacking around this like that it would be better to limit
the content of the paginator in the first place.